Kadrering als losse lagen met eigen checks, in de juiste volgorde
De kern van het probleem: de kadrering (gezichtsmeting, tracking, controle)
draaide voordat de knip- en aanloopcontrole de grenzen verschoven — daarna
sloeg alle meetdata nergens meer op. De pijplijn is nu een vaste laagvolgorde:

1. Grenzen definitief: knipcontrole, overlapregel, aanloopcorrectie. Elke
   latere grenswijziging meet het geraakte segment opnieuw.
2. Meten: spreker, gezichtsvak, spoor. Het spoor staat nu in absolute brontijd,
   zodat een grensverschuiving het volgen niet meer scheeftrekt.
3. Regel: een sprekerswissel is een knip, geen pan. Springt het spoor blijvend
   (>0,22 beeldbreedte), dan wordt het shot op het wisselmoment gesplitst in
   twee statische kaders — audio loopt door, beeld wisselt naar de spreker.
4. Rekenkundige kadercontrole plus visuele controle vooraf.
5. Na de render: eindbeeldcontrole op het eindbestand zelf — statische shots
   op het midden, bewegende shots op drie momenten. Fouten worden gecorrigeerd
   (ook op gevolgde shots: de duw gaat over het hele spoor heen) en er wordt
   een keer opnieuw gerenderd.

Schema van de visuele check accepteert nu ook gesplitste shotnummers (5.5).

Gemeten op de testclip: de eindbeeldcontrole ving twee kadreringsfouten in het
echte eindbestand, corrigeerde ze en de tweede render is schoon.

+++ src/lib/agents/kadercheck.ts
@@ -4,7 +4,7 @@ import { structuredCall } from '../claude';
 const checkSchema = z.object({
   shots: z.array(
     z.object({
-      volgorde: z.number().int().min(1),
+      volgorde: z.number().min(1),
       goed: z.boolean().describe('Staat de spreker correct in beeld?'),
       probleem: z
         .string()
